Add Android build to mobile app
Add Android build to mobile app

1. Android is added as a target platform. I executed the
follow command to do that:

ionic cordova run android --device

Most of the files in this commit are autogenerated by
this ionic command.

Althought we do not intend to publish our mobile apps to
Android in the first release I still added Android platform
so that I can test how the packaging and deployment works
with Ionic, and because Android was much simpler to set
than iOS. I will add iOS platform as a separate commit.

2. Added ionic CLI to locally installed node_modules

Ionic CLI can be now executed for example like this:

./node_modules/.bin/ionic serve

It is no longer required to install Ionic CLI globally.