Add linting rules to mobile app
The rules are based on angular-tslint-rules and codelyzer.
They define both TypeScript coding standards and Angular
style standards.

A few rules are overriden, see tslint.json

I had to touch a lot of files to make sure they confirm with
linting rules. Vast majority of changes should not affect
the functonality. The only functional change is removing
console.log calls. We need to replace them with some
centralized logging capability.

I also need to review "component-selector" and "trackBy-function"
rules to see if they are needed and enable them. They are
disabled for now.