Appointment status API
This commit extends existing retrieve/cancel endpoint of a single
appointment and adds ability to set it's status. Update scope is
currently limited only to the status field. Only limited set of
statuses are allowed to be set by a stylist.

Upon update, under the hood, timestamp and actor of the update are
recorded.

At later stages we may (or may not) want to extend this API and add
extra fields to be manually settable, e.g. some extra information
about checkout details, etc.