API support for multiple services and tax / card fee support
Changes to Appointment Create/Retrieve and for Appointment Preview endpoints.
From now on, instead of single service UUID, Appointment create and preview
endpoints will accept list of services uuids, signifying multiple services
to be included to the appointment.

Total duration and total cost are calculated as sum of all services. Tax and
card fee percentages are hard-coded for now.

Read-only fields `original_service_name` and `original_service_uuid` added
to retrive serializers to simplify one-line display of appointments on the
frontend and indicate with which service an appointment was originally
created.

Issue #99