Implement Invite Clients screen
- We send the invitation information to the backend but the
invitations are not actually sent to the phone numbers because
we don't yet have the Client App. The backend should be implemented
as soon as we release the Client App.

- Added generic alert creation function.

- Introduce a concept of incomplete feature flag via
ENV.ffEnableIncomplete flag. This is by default disabled on production.

TODO: the design and CSS must be cleaned up.

Github: #59