Automatically authenticate on startup and fix navigation
When the app loads we now verify if we have previusly save
authResponse. If we do then we will refresh the auth token
and show the correct screen based on the state of the profile.
If there is no valid auth saved then show the FirstScreen.

Also implemented a full history of navigation for the case
when after login we resume registration flow based on
completeness of the profile. In this case we will now also
set the navigation history in a way that ensures the Back
button works as expected.