Create shadow accounts for clients on appointment creation
If client is not supplied during appointment creation, we will create a
'shadow' account for the client, i.e. a Client model with User model
which is inactive, with unusable password and with auto-generated uuid-
based email at our domain.

Once the shadow account for the client is created, we will save the
client information to the appointment, and it will make possible for the
stylist to search for this client, and for us it will be possible to
track client's appointments and discounts.

Github issue: #